fix imports and cleanup
