+++ invenio_vocabularies/ext.py
@@ -44,10 +44,10 @@ from .contrib.subjects import (
 # from .contrib.information import InformationResource, InformationResourceConfig
 
 from .resources import (
-    VocabulariesResourceConfig,
     VocabularyTypeResourceConfig,
     VocabulariesResource,
     VocabulariesAdminResource,
+    VocabulariesResourceConfig,
 )
 from .services.service import VocabulariesService
 
@@ -129,10 +129,12 @@ class InvenioVocabularies(object):
             config=SubjectsResourceConfig,
         )
         self.resource = VocabulariesResource(
+            # connects resource with the config
             service=self.service,
             config=app.config["VOCABULARIES_RESOURCE_CONFIG"],
         )
         self.vocabulary_admin_resource = VocabulariesAdminResource(
+            # should connect the vocabulary_admin_resource with the config
             service=self.service,
             config=VocabularyTypeResourceConfig,
         )
