+++ invenio_vocabularies/resources/config.py
@@ -9,10 +9,25 @@
 """Resources config."""
 
 import marshmallow as ma
-from flask_resources import HTTPJSONException, ResourceConfig, create_error_handler
+from flask_resources import (
+    HTTPJSONException,
+    ResourceConfig,
+    create_error_handler,
+    JSONSerializer,
+    BaseListSchema,
+    MarshmallowSerializer,
+    ResponseHandler,
+)
+from invenio_records_resources.resources.records.headers import etag_headers
 from invenio_records_resources.resources.errors import ErrorHandlersMixin
 from invenio_records_resources.resources.records.args import SearchRequestArgsSchema
 from invenio_records_resources.services.base.config import ConfiguratorMixin
+from invenio_records_resources.resources import (
+    RecordResource,
+    RecordResourceConfig,
+    SearchRequestArgsSchema,
+)
+from .serializer import VocabularyL10NItemSchema
 
 
 class TasksResourceConfig(ResourceConfig, ConfiguratorMixin):
@@ -24,48 +39,52 @@ class TasksResourceConfig(ResourceConfig, ConfiguratorMixin):
     routes = {"list": ""}
 
 
-class VocabulariesSearchRequestArgsSchema(SearchRequestArgsSchema):
+class VocabularySearchRequestArgsSchema(SearchRequestArgsSchema):
     """Vocabularies search request parameters."""
 
+    tags = ma.fields.Str()
     active = ma.fields.Boolean()
-
-
-class VocabulariesResourceConfig(ResourceConfig, ConfiguratorMixin):
-    """Vocabularies resource config."""
-
-    # /vocabulary - all
-    # Blueprint configuration
-    blueprint_name = "vocabularies"
-    url_prefix = "/vocabularies"
-    routes = {
-        "list": "",
-        "item": "/<vocabulary_id>",
-    }
-
-    # Request parsing
-    request_read_args = {}
-    request_view_args = {"vocabulary_id": ma.fields.String}
-    request_search_args = VocabulariesSearchRequestArgsSchema
-
-    error_handlers = {
-        **ErrorHandlersMixin.error_handlers,
-        # TODO: Add custom error handlers here
-    }
-
-
-class VocabulariesSearchRequestArgsSchema(SearchRequestArgsSchema):
-    """Vocabularies search request parameters."""
-
     status = ma.fields.Boolean()
 
 
+class VocabulariesResourceConfig(RecordResourceConfig):
+    """Vocabulary resource configuration."""
+
+    blueprint_name = "vocabularies"
+    url_prefix = "/vocabularies"
+    routes = {
+        "list": "/<type>",
+        "item": "/<type>/<pid_value>",
+        "tasks": "/tasks",
+    }
+
+    request_view_args = {
+        "pid_value": ma.fields.Str(),
+        "type": ma.fields.Str(required=True),
+    }
+
+    request_search_args = VocabularySearchRequestArgsSchema
+
+    response_handlers = {
+        "application/json": ResponseHandler(JSONSerializer(), headers=etag_headers),
+        "application/vnd.inveniordm.v1+json": ResponseHandler(
+            MarshmallowSerializer(
+                format_serializer_cls=JSONSerializer,
+                object_schema_cls=VocabularyL10NItemSchema,
+                list_schema_cls=BaseListSchema,
+            ),
+            headers=etag_headers,
+        ),
+    }
+
+
 class VocabularyTypeResourceConfig(ResourceConfig, ConfiguratorMixin):
-    """Runs resource config."""
+    """Vocabulary list resource config."""
 
     # /vocabulary/vocabulary_id
     # Blueprint configuration
-    blueprint_name = "vocabulary_runs"
-    url_prefix = ""
+    blueprint_name = "vocabulary_list"
+    url_prefix = "/vocabularies"
 
     routes = {
         "all": "/",
@@ -74,12 +93,12 @@ class VocabularyTypeResourceConfig(ResourceConfig, ConfiguratorMixin):
     }
 
     # Request parsing
+    request_read_args = {}
     request_view_args = {
         "vocabulary_id": ma.fields.String,
         "vocabulary_type_id": ma.fields.String,
     }
-
-    request_search_args = VocabulariesSearchRequestArgsSchema
+    request_search_args = VocabularySearchRequestArgsSchema
 
     error_handlers = {
         **ErrorHandlersMixin.error_handlers,
