+++ setup.cfg
@@ -73,7 +73,7 @@ invenio_base.api_blueprints =
     invenio_vocabularies_names = invenio_vocabularies.views:create_names_blueprint_from_app
     invenio_vocabularies_subjects = invenio_vocabularies.views:create_subjects_blueprint_from_app
     invenio_vocabularies_ext = invenio_vocabularies.views:blueprint
-    invenio_vocabularies_list =  invenio_vocabularies.views:
+    invenio_vocabularies_list =  invenio_vocabularies.views:create_list_blueprint_from_app
 invenio_base.api_finalize_app =
     invenio_vocabularies = invenio_vocabularies.ext:api_finalize_app
 invenio_base.finalize_app =
